weekly-summary/pyproject.toml
[project]
name = "weekly-summary"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "wjdgks1224", email = "wjdgks1224@gmail.com" }
]
dependencies = [
    "gitpython>=3.1.43",
    "streamlit>=1.38.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"weekly-summary" = "weekly_summary:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/weekly_summary"]
